move elements into basics & components to mirror structure of hig.web
